Fixes
- Gave players more options to acquire coins: defeat enemies, sell towers and plants
- Fixed inability to place a tower that was occupied by a removed tower - a parameter
that determined whether or not a tower was occupying a tile was not being reset
- Fixed some quests not being updated - the statistics for them were not added to a list - created a statTable instead of a list
- Fixed upgraded towers doing inconsistent damage - bullets are pooled now, and damage is held in the bullets, so retrieving old bullets kept the previous damage and stacked the damage every time
- Fixed tower menu opening when placing a tower - made sure to check that the tower was placed before being able to open the menu

- Added a sell option to tower menu to acquire coins

+++ Assets/Menus/UpgradeMenu.cs
@@ -15,7 +15,7 @@ public class UpgradeMenu : MonoBehaviour
     public TMP_Text message;
 
     public float fireRateIncrease = 0.5f;
-    public float strengthIncrease = 1;
+    public int strengthIncrease = 1;
     public float rangeIncrease = 1;
 
     [Tooltip("Panel for insufficient funds")]
